move get_file_on_path from SPyVM to ImportAnalyzer
