+++ spy/vm/vm.py
@@ -4,7 +4,6 @@ from types import FunctionType
 from typing import Any, Callable, Iterable, Optional, Sequence, Union, overload
 
 import fixedint
-import py.path
 
 from spy import ROOT, ast, libspy
 from spy.analyze.symtable import Color, ImportRef, SymTable, maybe_blue
@@ -169,28 +168,6 @@ class SPyVM:
         importer.import_all()
         w_mod = self.modules_w[modname]
         return w_mod
-
-    def find_file_on_path(self, modname: str) -> Optional[py.path.local]:
-        # XXX for now we assume that we find the module as a single file in
-        # the only vm.path entry. Eventually we will need a proper import
-        # mechanism and support for packages
-        #
-        # We search dir by dir, and within a dir .spy wins over .py. The
-        # caller is expected to check the extension of the result: a .py file
-        # cannot be imported, but we return it anyway so that the caller can
-        # produce a good error message. Note that a .py found in an earlier
-        # dir shadows a .spy in a later dir.
-        assert self.path, "vm.path not set"
-        for d in self.path:
-            # XXX write test for this
-            f = py.path.local(d).join(f"{modname}.spy")
-            if f.exists():
-                return f
-            py_f = f.new(ext=".py")
-            if py_f.exists():
-                return py_f
-
-        return None
 
     def redshift(self, error_mode: ErrorMode) -> None:
         """
